analytics/requirements.txt
psycopg2-binary>=2.9
numpy>=1.26
scipy>=1.11
